dpdk: Windows support (part 2)

+++ drivers/net/liquidio/lio_rxtx.c
@@ -1426,7 +1426,7 @@ lio_delete_sglist(struct lio_instr_queue *txq)
 		if (g) {
 			if (g->sg)
 				rte_free(
-				    (void *)((unsigned long)g->sg - g->adjust));
+				    (void *)((uintptr_t)g->sg - g->adjust));
 			rte_free(g);
 		}
 	} while (g);
@@ -1471,10 +1471,10 @@ lio_setup_sglists(struct lio_device *lio_dev, int iq_no,
 		}
 
 		/* The gather component should be aligned on 64-bit boundary */
-		if (((unsigned long)g->sg) & 7) {
-			g->adjust = 8 - (((unsigned long)g->sg) & 7);
+		if (((uintptr_t)g->sg) & 7) {
+			g->adjust = 8 - (((uintptr_t)g->sg) & 7);
 			g->sg =
-			    (struct lio_sg_entry *)((unsigned long)g->sg +
+			    (struct lio_sg_entry *)((uintptr_t)g->sg +
 						       g->adjust);
 		}
